+++ pages/page-home.php
@@ -6,17 +6,7 @@
 <?php get_template_part('patterns/template-home-volumes'); ?>
 <?php get_template_part('patterns/template-home-categories'); ?>
 
-<div class="random-article">Non sai da dove partire? <a href="/?random=1">Lasciati ispirare</a> da uno dei molti articoli presenti.</div>
+<div class="random-article"><strong>Non sai da dove partire?</strong><br> <a href="/?random=1">Lasciati ispirare</a> da uno dei molti articoli presenti.</div>
 
-<section class="home-section" id="gift-of-dhamma">
-	<div style="text-align: center; margin: 4em 0 2em 0">
-		<?php echo do_shortcode( '[separator]' ) ?>
-	</div>
-	<div class="post-content">
-		<h2 class="wp-heading has-text-align-center">Il Dono del Dhamma</h2>
-
-		<p>Tutto il Dhamma qui esposto è offerto gratuitamente, senza pubblicità o donazioni suggerite. Questa è la diffusione del Dhamma secondo il Dhamma, in linea con quanto il Buddha insegnò riguardo a come i suoi insegnamenti dovessero essere offerti.</p>
-
-		<p>Per coloro che vogliono esprimere gratitudine per la generosità dei maestri, il modo migliore è mettere in pratica gli insegnamenti, con impegno sincero — per vedere con i propri occhi, per il proprio beneficio e per il beneficio di tutti gli esseri.</p>
-	</div>
-</section>
+<?php get_template_part('parts/dhamma-gift'); ?>
+<?php get_template_part('parts/singing-bowl'); ?>