+++ parts/cta-title-link.php
@@ -17,8 +17,10 @@ $title  = isset($args['title']) ? $args['title'] : '';
 $target = isset($args['target']) ? $args['target'] : '';
 $rel    = isset($args['rel'])    ? $args['rel']    : '';
 $id     = isset($args['id'])     ? $args['id']     : '';
+$desc   = isset($args['desc'])  ? $args['desc'] : '';
 $class  = 'cta-link' . (isset($args['class']) ? ' ' . esc_attr($args['class']) : '');
 ?>
+<div class="cta-link-container">
 <a
 	<?php if ($id) echo 'id="' . esc_attr($id) . '" '; ?>
 	class="<?php echo esc_attr($class); ?>"
@@ -30,3 +32,7 @@ $class  = 'cta-link' . (isset($args['class']) ? ' ' . esc_attr($args['class']) :
 
 	<svg class="cta-link-chevron" viewBox="0 0 24 24" width="48" height="48" aria-hidden="true"><path d="M6.23 8.97a1 1 0 0 1 1.41 0L12 13.34l4.36-4.37a1 1 0 1 1 1.41 1.42l-5.06 5.06a1 1 0 0 1-1.41 0L6.23 10.4a1 1 0 0 1 0-1.42z" fill="currentColor"></path></svg>
 </a>
+<?php if ($desc) : ?>
+    <p class="cta-link-desc"><? echo esc_html($desc); ?></p>
+<?php endif ?>
+</div>