+++ parts/svg/singing-bowl.php
@@ -1,0 +1,29 @@
+<svg viewBox="0 0 106.23385 54.588127" xmlns="http://www.w3.org/2000/svg" class="bowl-svg" role="img" aria-label="Bowl">
+  <g class="layer" transform="translate(-51.883076,-110.07287)">
+    <g transform="translate(8.339119)">
+      <!-- Outline / corpo -->
+      <path class="bowl-outline" d="m 147.94849,111.90218 a 51.287609,40.876385 0 0 1 -25.64381,35.39999 51.287609,40.876385 0 0 1 -51.287605,0 51.287609,40.876385 0 0 1 -25.643803,-35.39999 h 51.287609 z"/>
+      <!-- Base / cuscino -->
+      <path class="bowl-base" d="M 70.271243,162.89148 H 123.05052"/>
+      <!-- Incisioni -->
+      <g class="bowl-etches" transform="matrix(0.72754338,0,0,0.6180581,16.742591,45.893717)">
+        <path class="bowl-etch" d="m 55.504791,118.73774 5.370049,5.96672"/>
+        <path class="bowl-etch" d="m 65.811845,118.73774 5.370049,5.96672"/>
+        <path class="bowl-etch" d="m 76.118898,118.73774 5.370049,5.96672"/>
+        <path class="bowl-etch" d="m 86.425952,118.73774 5.370049,5.96672"/>
+        <path class="bowl-etch" d="m 96.733006,118.73774 5.370044,5.96672"/>
+        <path class="bowl-etch" d="m 107.04006,118.73774 5.37005,5.96672"/>
+        <path class="bowl-etch" d="m 117.34711,118.73774 5.37005,5.96672"/>
+        <path class="bowl-etch" d="m 127.65417,118.73774 5.37005,5.96672"/>
+        <path class="bowl-etch" d="m 137.49272,118.73774 5.37005,5.96672"/>
+        <path class="bowl-etch" d="m 147.33127,118.73774 5.37005,5.96672"/>
+        <path class="bowl-etch" d="m 157.16983,118.73774 5.37005,5.96672"/>
+      </g>
+    </g>
+  </g>
+  <!-- Stili basati su classi (identici agli originali) -->
+  <style>
+  .bowl-outline, .bowl-base, .bowl-etch { fill: none; stroke: #98989B; stroke-linecap: round; stroke-linejoin: round; } .bowl-outline { stroke-width: 3.65863; } .bowl-base { stroke-width: 3.53903; } .bowl-etch { stroke-width: 4; }
+    /* Se vuoi evitare che lo stroke cambi con le trasformazioni */ /* .bowl-etch { vector-effect: non-scaling-stroke; } */
+  </style>
+</svg>